add ability to enable and disable mag for BMI160
Todo: send magnetometer vector to the computer. The program lists this packet type as deprecated, so I don't know if this is possible.
The desktop app seems to have support to display it, but there's nothing in the core to take data from the tracker and throw it out in a format the frontend likes.

also moved old lsm6ds3 implementation to backups since the sensor fusion one should work and I don't have one to test with.